Refactor: artist to netartist
